pyproject.toml
[build-system]
build-backend = "poetry_dynamic_versioning.backend"
requires = [
  "poetry-core>=1",
  "poetry-dynamic-versioning",
]

[tool.poetry]
name = "tap-belvo"
version = "0.0.0"
description = "`tap-belvo` is a Singer tap for Belvo, built with the Meltano SDK for Singer Taps."
authors = ["Edgar Ramírez-Mondragón <edgarrm358@gmail.com>"]
keywords = [
  "ELT",
  "singer.io",
  "Belvo",
]
license = "Apache 2.0"
readme = "README.md"
homepage = "https://github.com/edgarrmondragon/tap-belvo"
repository = "https://github.com/edgarrmondragon/tap-belvo"
documentation = "https://github.com/edgarrmondragon/tap-belvo#readme"

[tool.poetry.dependencies]
python = ">=3.8,<4"
singer-sdk = "~=0.33.0"
requests-cache = "1.*"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
singer-sdk = { version = "~=0.33.0", extras = ["testing"] }

[tool.poetry.scripts]
# CLI declaration
"tap-belvo" = "tap_belvo.tap:TapBelvo.cli"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
use_parentheses = true
include_trailing_comma = true
src_paths = "tap_belvo"
known_first_party = "tap_belvo"
add_imports = [
  "from __future__ import annotations",
]

[tool.pytest.ini_options]
addopts = "-vvv"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = ["backoff.*"]

[tool.poetry-dynamic-versioning]
enable = true
format-jinja = """
    {%- if distance == 0 -%}
        {{ serialize_pep440(base, stage, revision) }}
    {%- elif revision is not none -%}
        {{ serialize_pep440(base, stage, revision + 1, dev=distance, metadata=[commit]) }}
    {%- else -%}
        {{ serialize_pep440(bump_version(base), stage, revision, dev=distance, metadata=[commit]) }}
    {%- endif -%}
"""
metadata = true
style = "pep440"
vcs = "git"

[tool.ruff]
ignore = [
  "ANN101",  # missing-type-self
  "DJ",      # flake8-django
  "FIX002",  # line-contains-todo
]
line-length = 88
select = ["ALL"]
src = ["tap_belvo", "tests"]
target-version = "py38"
unfixable = [
  "ERA001",  # commented-out-code
]

[tool.ruff.per-file-ignores]
"noxfile.py" = ["ANN"]
"tests/*" = [
  "ANN201",  # missing-return-type-public-function
  "S101",    # assert
  "SLF001",  # private-member-access
]

[tool.ruff.isort]
known-first-party = ["tap_belvo"]

[tool.ruff.pydocstyle]
convention = "google"
